README.txt
Python for Everybody is a 5 course Python specialization on python based on 5 major aspect
as follows:
  1. Python for everyone - Introduction and basics with graded Assignments
  2. Data-Structure - Retaining data temporarily with lists, tupels and Dicts
  3. Access Web Data - Socket and File Handle, Extraction of data from any type of website on XML & JSON format
  4. Database - Database degign, Relational DB, Data Modeling and Permanent Storage of data in python using SQLlite(python emmbedded) 
  5. Capstone - A recape of all you've Completed in the course basic focuse on ASCII and Data_structure
  
Course Instructor - DR Chuck Serverance
Host - University Of Mi0chigan
Platform - coursera

PERSONAL OPINION - This is my first Certified course ever taken so you should obviously take my words with a pile of salt...
Education for me has always been something forced upon with a lot of rules and how to do things, paths you have to follow.., play is next to faliure..,
all work no play makes jack a dull boy "But jack better be fucking smart to not play without reading all his books" and so on .... 
With Dr. Chuck it was different, I might be talking out of syntax here, but am sure if you just wanted to know about the course you probably wouldn't
be here. Dr. Chuck always believed in analogies and making fun facts out of everything he said, he made the class fun and convinient to attend.
If you have professional Software Knowledge on python already and you're not in dear need of a certificate this might not be the best course for you, 
but if your looking for how to get started on programming then this is the best course I have seen so far.!!